controller redirects and cleaning
